добавлен объектив 18-55

+++ Camera_Simulator/LenseInterface.h
@@ -12,8 +12,8 @@ public:
 						 size_t currentFocusDistanse, 
 						 size_t currentApperture) = 0;
 
-	virtual const double getMaxFocusDistanse() const = 0;
-	virtual const double getMinFocusDistanse() const = 0;
+	virtual const size_t getMaxFocusDistanse() const = 0;
+	virtual const size_t getMinFocusDistanse() const = 0;
 
 	virtual const double getMaxApperture() const = 0;
 	virtual const double getMinApperture() const = 0;
